dom/tests/mochitest/webcomponents/test_custom_element_runtime_exception.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1407669
-->
<head>
  <title>Test custom elements runtime exception</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1407669">Bug 1407669</a>
<script type="text/javascript">

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({set: [['dom.expose_test_interfaces', true]]}, function() {
  window.onerror = function (e) {
    ok(false, "How did we get here!?");
  }

  class Foo extends HTMLElement {
    constructor() {
      super()
      TestFunctions.throwUncatchableException();
    }
  }

  customElements.define("test-custom-element", Foo);
  document.createElement("test-custom-element");

  ok(true, "Uncatchable exception should not report");
  SimpleTest.finish();
});

</script>
</body>
</html>
